make openload update dynamically
